sixth commit

+++ src/main/java/tn/esb/lmad/flightPlannerAPI/Repositories/SeatRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import tn.esb.lmad.flightPlannerAPI.Domains.Seat;
 @Repository
-public interface SeatRepository extends JpaRepository<Seat,Short> {
+public interface SeatRepository extends JpaRepository<Seat,Long> {
 }
